+++ src/extension.ts
@@ -1,4 +1,4 @@
-import {EditorState, Transaction, ChangeSet, Facet, Extension} from "@codemirror/state"
+import {EditorState, Transaction, ChangeSet, Facet, StateEffect, Extension, SelectionRange} from "@codemirror/state"
 import {RangeSet} from "@codemirror/rangeset"
 import {StyleModule} from "style-mod"
 import {DecorationSet} from "./decoration"
@@ -27,6 +27,10 @@ export const exceptionSink = Facet.define<(exception: any) => void>()
 export const updateListener = Facet.define<(update: ViewUpdate) => void>()
 
 export const inputHandler = Facet.define<(view: EditorView, from: number, to: number, text: string) => boolean>()
+
+export const scrollTo = StateEffect.define<SelectionRange>({
+  map: (range, changes) => range.map(changes)
+})
 
 /// Log or report an unhandled exception in client code. Should
 /// probably only be used by extension code that allows client code to
@@ -272,7 +276,7 @@ export const decorations = Facet.define<DecorationSet>()
 
 export const styleModule = Facet.define<StyleModule>()
 
-export const enum UpdateFlag { Focus = 1, Height = 2, Viewport = 4, LineGaps = 8, Geometry = 16 }
+export const enum UpdateFlag { Focus = 1, Height = 2, Viewport = 4, Geometry = 8 }
 
 export class ChangedRange {
   constructor(readonly fromA: number, readonly toA: number, readonly fromB: number, readonly toB: number) {}
@@ -350,7 +354,9 @@ export class ViewUpdate {
     if (this.docChanged) this.flags |= UpdateFlag.Height
   }
 
-  /// Tells you whether the viewport changed in this update.
+  /// Tells you whether the [viewport](#view.EditorView.viewport) or
+  /// [visible ranges](#view.EditorView.visibleRanges) changed in this
+  /// update.
   get viewportChanged() {
     return (this.flags & UpdateFlag.Viewport) > 0
   }